Additional thoughts on WALSender cpu reduction. Use long type and alter a comment to reduce confusion.

+++ src/backend/replication/walsender.c
@@ -353,7 +353,7 @@ WalSndLoop(void)
 	/* Loop forever */
 	for (;;)
 	{
-		int			remain;		/* remaining time (ms) */
+		long	remain;		/* remaining time (us) */
 
 		/*
 		 * Emergency bailout if postmaster has died.  This is to avoid the
@@ -393,7 +393,7 @@ WalSndLoop(void)
 		 *
 		 * On some platforms, signals won't interrupt the sleep.  To ensure we
 		 * respond reasonably promptly when someone signals us, break down the
-		 * sleep into NAPTIME_PER_CYCLE (ms) increments, and check for
+		 * sleep into NAPTIME_PER_CYCLE increments, and check for
 		 * interrupts after each nap.
 		 */
 		remain = WalSndDelay * 1000L;
